refactor: move text box extension methods and tests to their own file
