+++ Velaptor/InternalExtensionMethods.cs
@@ -42,108 +42,7 @@ internal static class InternalExtensionMethods
         keyboardState.IsKeyDown(KeyCode.Up) ||
         keyboardState.IsKeyDown(KeyCode.Down);
 
-
-
-    public static void BumpAllToLeft(this List<(char character, NETRectF bounds)> charBounds, float amount)
-    {
-        for (var i = 0; i < charBounds.Count; i++)
-        {
-            var currItem = charBounds[i];
-            currItem.bounds.X -= amount;
-            charBounds[i] = currItem;
-        }
-    }
-
-    public static void BumpAllToRight(this List<(char character, NETRectF bounds)> charBounds, float amount)
-    {
-        for (var i = 0; i < charBounds.Count; i++)
-        {
-            var currItem = charBounds[i];
-            currItem.bounds.X += amount;
-            charBounds[i] = currItem;
-        }
-    }
-
-    public static void BumpAllUp(this List<(char character, NETRectF bounds)> charBounds, float amount)
-    {
-        for (var i = 0; i < charBounds.Count; i++)
-        {
-            var currItem = charBounds[i];
-            currItem.bounds.Y -= amount;
-            charBounds[i] = currItem;
-        }
-    }
-
-    public static void BumpAllDown(this List<(char character, NETRectF bounds)> charBounds, float amount)
-    {
-        for (var i = 0; i < charBounds.Count; i++)
-        {
-            var currItem = charBounds[i];
-            currItem.bounds.Y += amount;
-            charBounds[i] = currItem;
-        }
-    }
-
-    public static int TextLeft(this List<(char character, NETRectF bounds)> charBounds)
-    {
-        if (charBounds.Count <= 0)
-        {
-            return 0;
-        }
-
-        return (int)charBounds[0].bounds.Left;
-    }
-
-    public static int TextRight(this List<(char character, NETRectF bounds)> charBounds)
-    {
-        if (charBounds.Count <= 0)
-        {
-            return 0;
-        }
-
-        return (int)charBounds[^1].bounds.Right;
-    }
-
-    public static int TextWidth(this List<(char character, NETRectF bounds)> charBounds) => charBounds.TextRight() - charBounds.TextLeft();
-
-    public static float CharLeft(this List<(char character, NETRectF bounds)> charBounds, int index)
-    {
-        if (index < 0 || index >= charBounds.Count)
-        {
-            return 0f;
-        }
-
-        return charBounds[index].bounds.Left;
-    }
-
-    public static float CharRight(this List<(char character, NETRectF bounds)> charBounds, int index)
-    {
-        if (index < 0 || index >= charBounds.Count)
-        {
-            return 0f;
-        }
-
-        return charBounds[index].bounds.Right;
-    }
-
-    public static float CenterPositionX(this List<(char character, NETRectF bounds)>? charBounds)
-    {
-        if (charBounds is null || charBounds.Count <= 0)
-        {
-            return 0f;
-        }
-
-        var left = charBounds.Min(cb => cb.bounds.Left);
-        var right = charBounds.Max(cb => cb.bounds.Right);
-        var width = Math.Abs(left - right);
-
-        return left + width.Half();
-    }
-
-    public static bool GapAtRightEnd(this List<(char character, NETRectF bounds)>? charBounds, float rightEndLimitX) =>
-        charBounds is not null && charBounds.Count > 0 && charBounds.TextRight() < rightEndLimitX;
-
-  public static float Half(this float value) => value / 2;
+    public static float Half(this float value) => value / 2;
 
     public static int Half(this int value) => value / 2;
